Migrate docx export of checklist result
Now uses the new checklist filesystem layout, which is itself updated
to support the new behaviours.
Fixes ygb display.

Upgrades document download to use a temporary (in-memory) file and
Flask's send_file, rather than saving the document to a shared
location for downloading. This will prevent bugs when users try to
simultaneously download reports.

Existing problems with the report (in particular, newlines in the
content) are nto fixed in this change.